// src/app/_services/model/OnboardingGrid.model.ts
export class OnboardingGrid
{
     id: number;//ActionTransactionId
    ProcessTransactionId:number;
    ActionTransactionId: number;//
    ClientId:number;
    ClientName:string;
    ContractId:number;
    ContractName:string;
    MandateAssignmentId: number;
    Mandate:string;
    CandidateId: number;
    CandidateName: string;
    RequestedFor: string;
    ExpectedDOJ: Date;
    RequestedOn: Date;
    StatusId:number;
    Status: string;
    PendingAtUserId:number;
    PendingAtUserName:string;
    RejectedByUserId:number;
    RejectedByUserName: string;
    RejectionReason:string;
}